+++ persistence/account.go
@@ -5,6 +5,7 @@ import (
 	"math/big"
 
 	"github.com/pokt-network/pocket/persistence/types"
+	"github.com/pokt-network/pocket/shared/converters"
 
 	"github.com/jackc/pgx/v4"
 )
@@ -118,7 +119,9 @@ func (p PostgresContext) SubtractPoolAmount(name string, amount string) error {
 }
 
 // DISCUSS(team): If we are okay with `GetPoolAmount` return 0 as a default, this function can leverage
+//
 //	`operationPoolAmount` with `*orig = *delta` and make everything much simpler.
+//
 // DISCUSS(team): Do we have a use-case for this function?
 func (p PostgresContext) SetPoolAmount(name string, amount string) error {
 	ctx, tx, err := p.getCtxAndTx()
@@ -155,18 +158,18 @@ func (p *PostgresContext) operationPoolOrAccAmount(name, amount string,
 	if err != nil {
 		return err
 	}
-	originalAmountBig, err := types.StringToBigInt(originalAmount)
+	originalAmountBig, err := converters.StringToBigInt(originalAmount)
 	if err != nil {
 		return err
 	}
-	amountBig, err := types.StringToBigInt(amount)
+	amountBig, err := converters.StringToBigInt(amount)
 	if err != nil {
 		return err
 	}
 	if err := op(originalAmountBig, amountBig); err != nil {
 		return err
 	}
-	if _, err = tx.Exec(ctx, insert(name, types.BigIntToString(originalAmountBig), height)); err != nil {
+	if _, err = tx.Exec(ctx, insert(name, converters.BigIntToString(originalAmountBig), height)); err != nil {
 		return err
 	}
 	return nil